Add PERINTERVAL and PREVIOUS as reserved keywords
The PREVIOUS keyword provides a bit of an issue as a reserved keyword.
`previous` is already used as an expression for the `fill()` function.
When `previous` becomes a reserved keyword, this gets resolved as
something that isn't an ident anymore.

Some special handling has been added that causes a `previous` token in
the context of a expression to be evaluated as a special identity
literal (to avoid quoting), but this makes it a special exception among
the different special `fill()` parameters. It might just be better to
make the other `fill()` parameters reserved keywords too for simplicity.