Teach the parser about PERINTERVAL and PREVIOUS
When creating a continuous query, the parser now understands `PREVIOUS
int` and `PERINTERVAL int` options to customize the continuous query.

This change only modifies the parser to accept these options and make
them available when creating a continuous query. Further work still
needs to be done to have these options used.

This contributes to implementing issue #5136.